# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/movies": {
      "post": {
        "operationId": "MoviesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMovieDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a new movie",
        "tags": [
          "Movies"
        ]
      },
      "get": {
        "operationId": "MoviesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all movies",
        "tags": [
          "Movies"
        ]
      }
    },
    "/api/movies/{id}": {
      "get": {
        "operationId": "MoviesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a movie by ID",
        "tags": [
          "Movies"
        ]
      },
      "put": {
        "operationId": "MoviesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMovieDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a movie",
        "tags": [
          "Movies"
        ]
      },
      "patch": {
        "operationId": "MoviesController_updateMovie",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMovieDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Movies"
        ]
      },
      "delete": {
        "operationId": "MoviesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a movie",
        "tags": [
          "Movies"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Cine",
    "description": "api documentation",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateMovieDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Inception"
          },
          "year": {
            "type": "number",
            "example": 2010
          },
          "genre": {
            "type": "string",
            "example": "Sci-Fi"
          },
          "language": {
            "type": "string",
            "example": "English"
          },
          "duration": {
            "type": "number",
            "example": 148
          },
          "imageUrl": {
            "type": "string",
            "example": "https://image.url"
          },
          "trailerUrl": {
            "type": "string",
            "example": "https://trailer.url"
          },
          "description": {
            "type": "string",
            "example": "A mind-bending thriller"
          },
          "classification": {
            "type": "string",
            "example": "PG-13"
          },
          "format": {
            "example": [
              "2D",
              "3D"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "releaseDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-06-15"
          }
        },
        "required": [
          "title",
          "year",
          "genre",
          "language",
          "duration",
          "imageUrl",
          "trailerUrl",
          "description",
          "classification",
          "format",
          "releaseDate"
        ]
      },
      "UpdateMovieDto": {
        "type": "object",
        "properties": {

        }
      },
...
    }
  }
}